+++ src/main/java/raisetech/student/management/data/StudentsCourses.java
@@ -1,8 +1,7 @@
-package raisetech.student.management;
+package raisetech.student.management.data;
 
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
