+++ src/main/java/raisetech/student/management/service/StudentService.java
@@ -1,0 +1,27 @@
+package raisetech.student.management.service;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentsCourses;
+import raisetech.student.management.repository.StudentRepository;
+
+@Service
+public class StudentService {
+
+  private StudentRepository repository;
+
+  @Autowired
+  public StudentService(StudentRepository repository) {
+    this.repository = repository;
+  }
+
+  public List<Student> searchStudentList(){
+    return repository.searchStudent();
+  }
+
+  public List<StudentsCourses> searchStudentsCourseList(){
+    return repository.searchStudentsCourses();
+  }
+}
